refactor: テスト画面の画面幅判定とシャッフルを分離
